14: Increase maximum number of checkpoint accounts

+++ solana/programs/staking/src/context.rs
@@ -142,7 +142,7 @@ pub struct Delegate<'info> {
         _against_votes: u64,
         _for_votes: u64,
         _abstain_votes: u64,
-        stake_account_checkpoints_index: u8)]
+        stake_account_checkpoints_index: u16)]
 pub struct CastVote<'info> {
     #[account(mut)]
     pub owner: Signer<'info>,
@@ -429,7 +429,7 @@ pub struct CreateStakeAccount<'info> {
     // Stake program accounts:
     #[account(
         init,
-        seeds = [CHECKPOINT_DATA_SEED.as_bytes(), payer.key().as_ref(), 0u8.to_le_bytes().as_ref()],
+        seeds = [CHECKPOINT_DATA_SEED.as_bytes(), payer.key().as_ref(), 0u16.to_le_bytes().as_ref()],
         bump,
         payer = payer,
         space = checkpoints::CheckpointData::LEN,
